sort_levels_values function added.
Based on it we added the sort_cluster_numbers parameter to cutree. We are now able to have cutree return cluster id's which are ordered from left to right (so that the left most cluster in the tree will be "number 1").

Relevant tests are added.

This allows us to modify color_branches.