Added stepBothSides algorithm
Described in the paper 'Shuffle & untangle: novel untangle methods for solving the tanglegram layout problem' (Nguyen et al. 2022), the stepBothSides algorithm has yet to be implemented in a publicly repository to my knowledge. Here, I have added the option to use this untangling algorithm to the dendextend package in R