Added stepBothSides algorithm to untangle methods (#122)
* Added stepBothSides algorithm

Described in the paper 'Shuffle & untangle: novel untangle methods for solving the tanglegram layout problem' (Nguyen et al. 2022), the stepBothSides algorithm has yet to be implemented in a publicly repository to my knowledge. Here, I have added the option to use this untangling algorithm to the dendextend package in R

* Included 'stepBothSides' as option in method paramter for untangle.dendrogram

* added test

* Added self as contributor

* Added notes for update

* Added roxygen2 documentation

* compiled roxygen2 docs

---------

Co-authored-by: f006zvn1 <alec.buetow@dartmouth.edu>